Test cases, this thing works for odd sized arrays and arrays that are multiples of 2

+++ alblas.h
@@ -1,1 +1,1 @@
-void sscal(int N, float ALPHA, float *X, int INCX);
+void sscal(int N, float ALPHA, float *X, float INCX);
